backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1
celery==5.3.4
stripe==7.8.0
sendgrid==6.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
mypy==1.7.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
